requirements.txt
# Core dependencies
python-telegram-bot>=21.0
openai>=1.0
yt-dlp>=2024.0

# Optional - for analytics
SQLAlchemy>=2.0
psycopg2-binary>=2.9
